[Frontend] Implement addition of history event (grade/comment/file)

+++ cviewer_frontend/lib/domain/models/cv/cv_history_event_data.dart
@@ -13,5 +13,5 @@ class CVHistoryEventData {
   final int authorId;
   final FileInfo? fileInfo;
   final String? comment;
-  final double? grade;
+  final int? grade;
 }
